Adds validation script that uses Melanie's read simulation.
The reads Melanie simulated have the name of the reference they are from in
their name. After mapping the simulated reads, one can therefore determine
where each read is stemming from. This script outputs a table where each row
has the counts per genome for a contig, which can be useful for validation.